requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
coverage==6.4.1
dj-database-url==0.5.0
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
flake8==4.0.1
gunicorn==20.1.0
isort==5.10.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
whitenoise==6.2.0
